fix: "/" added after "baseURL" in configuration
The base URL for localhost turns out to have a forward slash ("/") at
the end, because of which no "/" was added after the "baseURL" value in
"layouts/partials/style.html" while adding a custom CSS file. But since
the "baseURL" for production did not have a "/" at the end, the path for
the custom CSS file was being built incorrectly (".../blogcss/custom.css"
instead of ".../blog.css/custom.css").

This has been fixed by adding a "/" at the end of "baseURL".